Add comprehensive IADS and SEAD simulation capabilities
- Implement Integrated Air Defense System (IADS) with C4I network
- Add multi-layer SAM systems (long/medium/short range)
- Create Wild Weasel SEAD aircraft with AGM-88 HARM
- Implement radar warning receiver and emitter location
- Add EMCON tactics and memory mode for shutdown counters
- Create engagement simulations and visualizations

Key features:
- Layered defense with strategic, tactical, and point defense SAMs
- Realistic missile kinematics and proportional navigation
- RWR detection with triangulation for emitter location
- HARM anti-radiation missile with multiple operating modes
- SAM site EMCON management to counter SEAD
- Track-to-track correlation in IADS network
- Threat assessment and weapon assignment logic

SAM capabilities:
- Long-range (400km): S-400/Patriot equivalent
- Medium-range (50km): Buk/NASAMS equivalent
- Short-range (20km): Pantsir/C-RAM equivalent
- Realistic engagement envelopes and Pk calculations

SEAD capabilities:
- F-16CJ Wild Weasel simulation
- AGM-88 HARM with POS, HAS, TOO, and EOM modes
- RWR with bearing accuracy and signal strength
- Multiple SEAD tactics (preemptive, reactive, standoff)
- Memory mode for engaging shutdown emitters

+++ src/networked_radar/distributed_tracking.py
@@ -222,14 +222,14 @@ class DistributedTrackFusion:
             # Optimize for multiple tracks
             weights = self._optimal_omega_multi_tracks([t.covariance for t in tracks])
         
-        # Perform fusion
-        P_fused_inv = np.zeros_like(tracks[0].covariance)
-        x_fused_weighted = np.zeros_like(tracks[0].state)
+        # Perform fusion - ensure float64 dtype
+        P_fused_inv = np.zeros_like(tracks[0].covariance, dtype=np.float64)
+        x_fused_weighted = np.zeros_like(tracks[0].state, dtype=np.float64)
         
         for i, track in enumerate(tracks):
-            P_inv = np.linalg.inv(track.covariance)
+            P_inv = np.linalg.inv(track.covariance.astype(np.float64))
             P_fused_inv += weights[i] * P_inv
-            x_fused_weighted += weights[i] * P_inv @ track.state
+            x_fused_weighted += weights[i] * P_inv @ track.state.astype(np.float64)
         
         P_fused = np.linalg.inv(P_fused_inv)
         x_fused = P_fused @ x_fused_weighted
@@ -273,13 +273,13 @@ class DistributedTrackFusion:
         if len(tracks) == 1:
             return tracks[0]
         
-        # Convert to information form
-        Y_total = np.zeros_like(tracks[0].covariance)
-        y_total = np.zeros_like(tracks[0].state)
+        # Convert to information form - ensure float64 dtype
+        Y_total = np.zeros_like(tracks[0].covariance, dtype=np.float64)
+        y_total = np.zeros_like(tracks[0].state, dtype=np.float64)
         
         for track in tracks:
-            Y = np.linalg.inv(track.covariance)  # Information matrix
-            y = Y @ track.state  # Information vector
+            Y = np.linalg.inv(track.covariance.astype(np.float64))  # Information matrix
+            y = Y @ track.state.astype(np.float64)  # Information vector
             Y_total += Y
             y_total += y
